+++ internal/cli/app.go
@@ -1,26 +1,30 @@
 package cli
 
 import (
+	"context"
 	"fmt"
 	"runtime"
 	"strings"
 
 	"github.com/urfave/cli/v2"
 
+	"github.com/tarampampam/error-pages/internal/checkers"
 	"github.com/tarampampam/error-pages/internal/cli/build"
+	"github.com/tarampampam/error-pages/internal/cli/healthcheck"
+	"github.com/tarampampam/error-pages/internal/cli/serve"
 	"github.com/tarampampam/error-pages/internal/env"
 	"github.com/tarampampam/error-pages/internal/logger"
 	"github.com/tarampampam/error-pages/internal/version"
 )
 
 // NewApp creates new console application.
-func NewApp(appName string) *cli.App {
+func NewApp(appName string) *cli.App { //nolint:funlen
 	const (
 		logLevelFlagName  = "log-level"
 		logFormatFlagName = "log-format"
 		verboseFlagName   = "verbose"
 		debugFlagName     = "debug"
-		logJsonFlagName   = "log-json"
+		logJSONFlagName   = "log-json"
 
 		defaultLogLevel  = logger.InfoLevel
 		defaultLogFormat = logger.ConsoleFormat
@@ -45,7 +49,7 @@ func NewApp(appName string) *cli.App {
 				}
 			}
 
-			if c.Bool(logJsonFlagName) {
+			if c.Bool(logJSONFlagName) {
 				logFormat = logger.JSONFormat
 			} else {
 				// parse logging format
@@ -64,24 +68,23 @@ func NewApp(appName string) *cli.App {
 			return nil
 		},
 		Commands: []*cli.Command{
-			// versionCmd.NewCommand(version.Version()),
-			// healthcheckCmd.NewCommand(checkers.NewHealthChecker(ctx)),
+			healthcheck.NewCommand(checkers.NewHealthChecker(context.TODO())),
 			build.NewCommand(log),
-			// serveCmd.NewCommand(ctx, log, &configFile),
+			serve.NewCommand(log),
 		},
 		Version: fmt.Sprintf("%s (%s)", version.Version(), runtime.Version()),
 		Flags: []cli.Flag{ // global flags
 			&cli.BoolFlag{ // kept for backward compatibility
 				Name:  verboseFlagName,
-				Usage: "verbose output",
+				Usage: "verbose output (DEPRECATED FLAG)",
 			},
 			&cli.BoolFlag{ // kept for backward compatibility
 				Name:  debugFlagName,
-				Usage: "debug output",
+				Usage: "debug output (DEPRECATED FLAG)",
 			},
 			&cli.BoolFlag{ // kept for backward compatibility
-				Name:  logJsonFlagName,
-				Usage: "logs in JSON format",
+				Name:  logJSONFlagName,
+				Usage: "logs in JSON format (DEPRECATED FLAG)",
 			},
 			&cli.StringFlag{
 				Name:    logLevelFlagName,